Add basics of a `--parallel` switch for dicon.deploy
This allows for a `--parallel` switch for `dicon.deploy` which will
deploy to hosts in parallel. Since this deployment to multiple hosts in
parallel will also require capturing IO, there's also a `silent`
argument that's been added to a lot of things for if it should run
silently.

It's pretty bare bones, and there are no tests at this point, but it
basically works.